+++ example/lib/counter/show_add_button.dart
@@ -1,62 +1,78 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:example/counter/counter.model.dart';
+import 'package:iconnect/iconnect.dart';
 
 class ShowButtonOrigin extends StatelessWidget {
-  ShowButtonOrigin({Key? key, this.increase}) : super(key: key);
+  const ShowButtonOrigin({Key? key, this.increase}) : super(key: key);
   final VoidCallback? increase;
   @override
   Widget build(BuildContext context) {
     return FloatingActionButton(
       onPressed: increase,
       tooltip: 'Increment',
-      child: Icon(Icons.add),
+      child: const Icon(Icons.add),
     );
   }
 }
 
 class ShowButton extends StatelessWidget {
+  const ShowButton({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
-    return ShowButtonOrigin(
-        increase: () => first.dispatch(() => first.increment(1)));
+    //print('press button');
+    return ShowButtonOrigin(increase: () {
+      first.increment(1);
+      dispatch<CounterModel>();
+    });
   }
 }
 
 class ShowButtonSecond extends StatelessWidget {
+  const ShowButtonSecond({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
-    print('ShowButtonSecond build');
-    return ShowButtonOrigin(
-        increase: () => second.dispatch(() => second.increment(1)));
+    if (kDebugMode) {
+      print('ShowButtonSecond build');
+    }
+    return ShowButtonOrigin(increase: () {
+      second.increment(1);
+      dispatch<CounterModel>('second');
+    });
   }
 }
 
 class ShowButtonAsync extends StatelessWidget {
+  const ShowButtonAsync({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return ShowButtonOrigin(increase: () async {
-      third.snapshot = AsyncSnapshot.waiting();
-
-      third.dispatch();
+      third.snapshot = const AsyncSnapshot.waiting();
+      dispatch<CounterModel>('third');
       try {
         await third.incrementAsync(1);
 
         third.snapshot = AsyncSnapshot.withData(ConnectionState.done, third);
-        third.dispatch();
+        dispatch<CounterModel>('third');
       } catch (e) {
         third.snapshot = AsyncSnapshot.withError(ConnectionState.done, e);
-        third.dispatch();
+        dispatch<CounterModel>('third');
       } finally {}
     });
   }
 }
 
 class ShowButtonStream extends StatelessWidget {
+  const ShowButtonStream({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return ShowButtonOrigin(increase: () {
       fourth.streamIncrease().listen((event) {
-        fourth.dispatch();
+        dispatch<CounterModel>('fourth');
       });
     });
   }