+++ example/lib/counter/show_title.dart
@@ -1,13 +1,16 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 class ShowTitle extends StatelessWidget {
-  ShowTitle(
+  const ShowTitle(
       {this.title = 'You have pushed the button this many times:', Key? key})
       : super(key: key);
   final String title;
   @override
   Widget build(BuildContext context) {
-    print('ShowTitle build');
+    if (kDebugMode) {
+      print('ShowTitle build');
+    }
     return Text(title);
   }
 }