+++ lib/iconnect.dart
@@ -1,12 +1,13 @@
+import 'package:flutter/material.dart';
+
 import 'src/store.dart';
-export 'src/mixin.dart';
+//export 'src/mixin.dart';
 
 //two global function:provider and connect
 //one mixin for model:IConnect
-get provider => Store.instance.provider;
-
+Widget Function(Widget) get provider => Store.instance.provider;
 //if don't define model with mixin IConnect,could use these four function,for register,listen,dispatch
-get listen => Store.instance.listen;
-get dispatch => Store.instance.dispatch;
-get register => Store.instance.register;
-get unregister => Store.instance.unRegister;
+Function<T>(BuildContext, [String?]) get listen => Store.instance.listen;
+void Function<T>([String? key]) get dispatch => Store.instance.dispatch;
+T Function<T>(T, [String? key]) get register => Store.instance.register;
+void Function<T>([String?]) get unregister => Store.instance.unRegister;