+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.4.0 _2022.10.3
+1. dispatch now used without action
+1. remove dipose function,try dipose if exist in model
+1. use _rebuild instead of setState
+1. remove mixin 
+1. return future.error in model to future builder
+1. use Type as key in map,instead instance
+1. support multi instance,by string key
+
 ## 0.3.3 - 2021.11.10
 1. remove iconnect,use Builder widget instead
 1. fix bug in example,register should have an argument